services/tg-bot/requirements.txt
python-telegram-bot
httpx
psycopg2-binary
